mavsdk_server: widen the system ID API to 32 bits
mavsdk_server_run_with_mavlink_ids() is used by the language wrappers,
so widen its system ID now as well, to avoid another API break once 32
bit system IDs are implemented.

The --sysid argument of the binary still only accepts 1..255.

+++ cpp/src/mavsdk_server/src/mavsdk_server_api.cpp
@@ -28,7 +28,7 @@ int mavsdk_server_run_with_mavlink_ids(
     MavsdkServer* mavsdk_server,
     const char* system_address,
     const int mavsdk_server_port,
-    const uint8_t system_id,
+    const uint32_t system_id,
     const uint8_t component_id)
 {
     mavsdk_server->setMavlinkIds(system_id, component_id);
